requirements.txt
faster-whisper>=1.0.3
sounddevice>=0.4.6
numpy>=1.24
scipy>=1.10
keyboard>=0.13.5
pyperclip>=1.8.2
pyautogui>=0.9.54
pystray>=0.19.5
Pillow>=10.0.0

# GPU/CUDA-stod (valfritt, stor nedladdning ~2.5GB):
# pip install torch --index-url https://download.pytorch.org/whl/cu124
